2nd Pass

+++ src/sx128x.h
@@ -102,7 +102,7 @@ class Sx128xDriverBase
     // Rx methods
 
     void SetRx(uint8_t PeriodBase, uint16_t PeriodBaseCount);
-    void GetPacketStatus(int8_t* RssiSync, int8_t* Snr);
+    void GetPacketStatus(int16_t* RssiSync, int8_t* Snr);
     void GetRxBufferStatus(uint8_t* rxPayloadLength, uint8_t* rxStartBufferPointer);
 
     // auxiliary methods
